Move base client into own ducks file
